ARCO-220: introduce background worker to periodically check gaps in blocks
