+++ internal/blocktx/background_workers.go
@@ -1,0 +1,108 @@
+package blocktx
+
+import (
+	"context"
+	"log/slog"
+	"sync"
+	"time"
+
+	"github.com/bitcoin-sv/arc/internal/blocktx/store"
+	"github.com/libsv/go-p2p"
+)
+
+type BackgroundWorkers struct {
+	l *slog.Logger
+	s store.BlocktxStore
+
+	workersWg sync.WaitGroup
+	ctx       context.Context
+	cancelAll func()
+}
+
+func NewBackgroundWorkers(store store.BlocktxStore, logger *slog.Logger) *BackgroundWorkers {
+	ctx, cancel := context.WithCancel(context.Background())
+
+	return &BackgroundWorkers{
+		s: store,
+		l: logger.With(slog.String("module", "background workers")),
+
+		ctx:       ctx,
+		cancelAll: cancel,
+	}
+}
+
+func (w *BackgroundWorkers) GracefulStop() {
+	w.l.Info("Shutting down")
+
+	w.cancelAll()
+	w.workersWg.Wait()
+
+	w.l.Info("Shutted down")
+}
+
+func (w *BackgroundWorkers) StartFillGaps(peers []p2p.PeerI, interval time.Duration, retentionDays int, blockRequestCh chan<- BlockRequest) {
+	w.workersWg.Add(1)
+
+	go func() {
+		t := time.NewTicker(interval)
+		peerIndex := 0
+
+		for {
+			select {
+			case <-t.C:
+				if peerIndex >= len(peers) {
+					peerIndex = 0
+				}
+
+				err := w.fillGaps(peers[peerIndex], retentionDays, blockRequestCh)
+				if err != nil {
+					w.l.Error("failed to fill blocks gaps", slog.String("err", err.Error()))
+				}
+
+				peerIndex++
+				t.Reset(interval)
+
+			case <-w.ctx.Done():
+				w.workersWg.Done()
+				return
+			}
+
+		}
+	}()
+}
+
+func (w *BackgroundWorkers) fillGaps(peer p2p.PeerI, retentionDays int, blockRequestCh chan<- BlockRequest) error {
+	const (
+		hoursPerDay   = 24
+		blocksPerHour = 6
+	)
+
+	heightRange := retentionDays * hoursPerDay * blocksPerHour
+	blockHeightGaps, err := w.s.GetBlockGaps(w.ctx, heightRange)
+	if err != nil {
+		return err
+	}
+
+	if len(blockHeightGaps) == 0 {
+		return nil
+	}
+
+	for i, block := range blockHeightGaps {
+		if i == maxRequestBlocks {
+			break
+		}
+
+		w.l.Info("adding request for missing block to request channel",
+			slog.String("hash", block.Hash.String()),
+			slog.Uint64("height", block.Height),
+			slog.String("peer", peer.String()),
+		)
+
+		blockRequestCh <- BlockRequest{
+			Hash: block.Hash,
+			Peer: peer,
+		}
+	}
+
+	return nil
+}
