+++ internal/blocktx/processor_test.go
@@ -409,102 +409,6 @@ func TestHandleBlockReorg(t *testing.T) {
 	}
 }
 
-func TestStartFillGaps(t *testing.T) {
-	hostname, err := os.Hostname()
-	require.NoError(t, err)
-
-	tt := []struct {
-		name            string
-		hostname        string
-		getBlockGapsErr error
-		blockGaps       []*store.BlockGap
-
-		minExpectedGetBlockCapsCalls int
-	}{
-		{
-			name:     "success",
-			hostname: hostname,
-			blockGaps: []*store.BlockGap{
-				{
-					Height: 822014,
-					Hash:   testdata.Block1Hash,
-				},
-				{
-					Height: 822015,
-					Hash:   testdata.Block2Hash,
-				},
-			},
-
-			minExpectedGetBlockCapsCalls: 1,
-		},
-		{
-			name:            "error getting block gaps",
-			hostname:        hostname,
-			getBlockGapsErr: errors.New("failed to get block gaps"),
-
-			minExpectedGetBlockCapsCalls: 1,
-		},
-		{
-			name:      "no block gaps",
-			hostname:  hostname,
-			blockGaps: make([]*store.BlockGap, 0),
-
-			minExpectedGetBlockCapsCalls: 4,
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			// given
-			getBlockErrCh := make(chan error)
-
-			getBlockGapTestErr := tc.getBlockGapsErr
-			storeMock := &storeMocks.BlocktxStoreMock{
-				GetBlockGapsFunc: func(ctx context.Context, heightRange int) ([]*store.BlockGap, error) {
-					if getBlockGapTestErr != nil {
-						getBlockErrCh <- getBlockGapTestErr
-						return nil, getBlockGapTestErr
-					}
-
-					return tc.blockGaps, nil
-				},
-			}
-
-			logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo}))
-
-			blockRequestCh := make(chan blocktx.BlockRequest, 10)
-			var blockProcessCh chan *p2p.BlockMessage = nil
-
-			// when
-			_ = blocktx.NewPeerHandler(logger, blockRequestCh, blockProcessCh)
-			sut, err := blocktx.NewProcessor(logger, storeMock, blockRequestCh, blockProcessCh, blocktx.WithFillGapsInterval(time.Millisecond*20))
-			require.NoError(t, err)
-
-			peerMock := &mocks.PeerMock{
-				StringFunc: func() string {
-					return ""
-				},
-			}
-			peers := []p2p.PeerI{peerMock}
-
-			sut.StartFillGaps(peers)
-
-			select {
-			case hashPeer := <-sut.GetBlockRequestCh():
-				require.True(t, testdata.Block1Hash.IsEqual(hashPeer.Hash))
-			case err = <-getBlockErrCh:
-				require.ErrorIs(t, err, tc.getBlockGapsErr)
-			case <-time.NewTimer(100 * time.Millisecond).C:
-			}
-
-			sut.Shutdown()
-
-			// then
-			require.GreaterOrEqual(t, len(storeMock.GetBlockGapsCalls()), tc.minExpectedGetBlockCapsCalls)
-		})
-	}
-}
-
 func TestStartProcessRegisterTxs(t *testing.T) {
 	tt := []struct {
 		name        string
